+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 
 ## 0.27.8 – Aug 12, 2020
 
+### New
+- `mam` mutation `dropCachedDbInfo` to reset cached indexes.
+ 
 ### Fix
 - Update indexes retries if index creation has failed on the timeout.